Footer pushed down to bottom using flexbox
Made it so that all content is part of a flexbox, which has been made to push the footer to the bottom. Currently the header is not part of that flexbox which I think is leading to the footer being pushed down a bit further, but I'm happy with how it is for now.